+++ tests/testthat/test-do_GroupwiseDEHeatmap.R
@@ -1,5 +1,5 @@
-if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
-  testthat::test_that("do_GroupwiseDEPlot: CRAN essentials", {
+if(base::isFALSE(dep_check[["do_GroupwiseDEHeatmap"]])){
+  testthat::test_that("do_GroupwiseDEHeatmap: CRAN essentials", {
     suppressWarnings({
     sample <- SeuratObject::SetAssayData(object = sample,
                                          assay = "SCT",
@@ -9,20 +9,20 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                                                                          slot = "data")))
     })
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data")
     testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes_scaled,
                                     assay = "SCT",
                                     slot = "scale.data")
     testthat::expect_type(p, "list")
   })
 
-  testthat::test_that("do_GroupwiseDEPlot: PASS - default", {
+  testthat::test_that("do_GroupwiseDEHeatmap: PASS - default", {
     testthat::skip_on_cran()
 
 
@@ -35,7 +35,7 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                                                                        slot = "data")))
     })
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
@@ -43,7 +43,7 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                     sequential.direction = 1)
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
@@ -51,7 +51,7 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                     sequential.direction = -1)
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
@@ -59,7 +59,7 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                     viridis.direction = 1)
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
@@ -67,14 +67,14 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                     viridis.direction = -1)
     testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes_scaled,
                                     assay = "SCT",
                                     slot = "scale.data")
     testthat::expect_type(p, "list")
 
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
@@ -83,7 +83,7 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                     min.cutoff = 1)
     testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
@@ -91,7 +91,7 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                     min.cutoff = 1)
     testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
@@ -101,7 +101,7 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
 
   })
 
-  testthat::test_that("do_GroupwiseDEPlot: PASS - heatmap legend side", {
+  testthat::test_that("do_GroupwiseDEHeatmap: PASS - heatmap legend side", {
     testthat::skip_on_cran()
 
     suppressWarnings({
@@ -113,14 +113,14 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
                                                                                        slot = "data")))
     })
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes,
                                     assay = "SCT",
                                     slot = "data",
                                     legend.position = "right")
     testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
+    p <- SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                     de_genes = de_genes_scaled,
                                     assay = "SCT",
                                     slot = "scale.data",
@@ -129,10 +129,10 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
   })
 
 
-  testthat::test_that("do_GroupwiseDEPlot: FAIL - wrong direction", {
+  testthat::test_that("do_GroupwiseDEHeatmap: FAIL - wrong direction", {
     testthat::skip_on_cran()
 
-    testthat::expect_error({SCpubr::do_GroupwiseDEPlot(sample = sample,
+    testthat::expect_error({SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                                        de_genes = de_genes,
                                                        assay = "SCT",
                                                        slot = "data",
@@ -141,21 +141,21 @@ if(base::isFALSE(dep_check[["do_GroupwiseDEPlot"]])){
 
   testthat::test_that("do_ExpressionHeatmap: FAIL", {
     testthat::skip_on_cran()
-    testthat::expect_error({SCpubr::do_GroupwiseDEPlot(sample = sample,
+    testthat::expect_error({SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                                        de_genes = de_genes,
                                                        assay = "SCT",
                                                        slot = "data",
                                                        viridis.direction = 1,
                                                        min.cutoff = -10)})
 
-    testthat::expect_error({SCpubr::do_GroupwiseDEPlot(sample = sample,
+    testthat::expect_error({SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                                        de_genes = de_genes,
                                                        assay = "SCT",
                                                        slot = "data",
                                                        viridis.direction = 1,
                                                        max.cutoff = 200)})
 
-    testthat::expect_error({SCpubr::do_GroupwiseDEPlot(sample = sample,
+    testthat::expect_error({SCpubr::do_GroupwiseDEHeatmap(sample = sample,
                                                        de_genes = de_genes,
                                                        assay = "SCT",
                                                        slot = "data",